Moved some of code transform
