--- lib/Application/Exception/FileAlreadyExists.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Phpactor\Application\Exception;
-
-class FileAlreadyExists extends \Exception
-{
-}
