+++ lib/Container/RpcExtension.php
@@ -14,9 +14,9 @@ use Phpactor\Extension\ClassMover\Rpc\ClassCopyHandler;
 use Phpactor\Extension\ClassMover\Rpc\ClassMoveHandler;
 use Phpactor\Extension\ClassMover\Rpc\ReferencesHandler;
 use Phpactor\Rpc\Handler\OffsetInfoHandler;
-use Phpactor\Rpc\Handler\TransformHandler;
-use Phpactor\Rpc\Handler\ClassNewHandler;
-use Phpactor\Rpc\Handler\ClassInflectHandler;
+use Phpactor\Extension\CodeTransform\Rpc\TransformHandler;
+use Phpactor\Extension\CodeTransform\Rpc\ClassNewHandler;
+use Phpactor\Extension\CodeTransform\Rpc\ClassInflectHandler;
 use Phpactor\Rpc\Handler\ContextMenuHandler;
 use Phpactor\Rpc\Handler\ExtractConstantHandler;
 use Phpactor\Rpc\Handler\ExtractMethodHandler;
