+++ lib/Extension/CodeTransform/Application/Exception/FileAlreadyExists.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Phpactor\Extension\CodeTransform\Application\Exception;
+
+class FileAlreadyExists extends \Exception
+{
+}
