+++ lib/Extension/CodeTransform/CodeTransformExtension.php
@@ -7,8 +7,8 @@ use Phpactor\CodeTransform\Adapter\WorseReflection\Transformer\ImplementContract
 use Phpactor\CodeTransform\CodeTransform;
 use Phpactor\CodeTransform\Domain\Generators;
 use Phpactor\CodeTransform\Domain\Transformers;
-use Phpactor\Console\Command\ClassNewCommand;
-use Phpactor\Console\Command\ClassTransformCommand;
+use Phpactor\Extension\CodeTransform\Command\ClassNewCommand;
+use Phpactor\Extension\CodeTransform\Command\ClassTransformCommand;
 use Phpactor\CodeTransform\Adapter\Native\GenerateNew\ClassGenerator;
 use Phpactor\CodeBuilder\Adapter\Twig\TwigRenderer;
 use Phpactor\Extension\CodeTransform\Application\ClassNew;
@@ -17,7 +17,7 @@ use Twig\Environment;
 use Phpactor\CodeBuilder\Adapter\Twig\TwigExtension;
 use Twig\Loader\ChainLoader;
 use Phpactor\CodeTransform\Adapter\WorseReflection\GenerateFromExisting\InterfaceFromExistingGenerator;
-use Phpactor\Console\Command\ClassInflectCommand;
+use Phpactor\Extension\CodeTransform\Command\ClassInflectCommand;
 use Phpactor\Extension\CodeTransform\Application\ClassInflect;
 use Phpactor\CodeBuilder\Domain\Updater;
 use Phpactor\CodeBuilder\Adapter\TolerantParser\TolerantUpdater;
