+++ lib/Extension/CodeTransform/Command/ClassNewCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Console\Command;
+namespace Phpactor\Extension\CodeTransform\Command;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -11,8 +11,9 @@ use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Question\Question;
 use Phpactor\Extension\CodeTransform\Application\ClassNew;
 use Phpactor\Console\Dumper\DumperRegistry;
-use Phpactor\Application\Exception\FileAlreadyExists;
+use Phpactor\Extension\CodeTransform\Application\Exception\FileAlreadyExists;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
+use Phpactor\Console\Command\Handler\FormatHandler;
 
 class ClassNewCommand extends Command
 {
@@ -43,7 +44,7 @@ class ClassNewCommand extends Command
         $this->addOption('variant', null, InputOption::VALUE_REQUIRED, 'Variant', 'default');
         $this->addOption('list', null, InputOption::VALUE_NONE, 'List variants');
         $this->addOption('force', null, InputOption::VALUE_NONE, 'Force overwriting');
-        Handler\FormatHandler::configure($this);
+        FormatHandler::configure($this);
     }
 
     public function execute(InputInterface $input, OutputInterface $output)
