+++ lib/Extension/CodeTransform/Rpc/ClassInflectHandler.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Phpactor\Rpc\Handler;
+namespace Phpactor\Extension\CodeTransform\Rpc;
 
 use Phpactor\Rpc\Handler;
 use Webmozart\Glob\Glob;
+use Phpactor\Extension\CodeTransform\Rpc\AbstractClassGenerateHandler;
 
 class ClassInflectHandler extends AbstractClassGenerateHandler
 {
