+++ lib/Extension/CodeTransform/Rpc/ClassNewHandler.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Phpactor\Rpc\Handler;
+namespace Phpactor\Extension\CodeTransform\Rpc;
+
+use Phpactor\Extension\CodeTransform\Rpc\AbstractClassGenerateHandler;
 
 class ClassNewHandler extends AbstractClassGenerateHandler
 {
