+++ tests/Unit/Rpc/Handler/TransformHandlerTest.php
@@ -7,7 +7,7 @@ use Phpactor\CodeTransform\Domain\Transformer;
 use Phpactor\Rpc\Response\Input\ChoiceInput;
 use Phpactor\Rpc\Response\InputCallbackResponse;
 use Phpactor\CodeTransform\CodeTransform;
-use Phpactor\Rpc\Handler\TransformHandler;
+use Phpactor\Extension\CodeTransform\Rpc\TransformHandler;
 use Phpactor\CodeTransform\Domain\Transformers;
 use Phpactor\CodeTransform\Domain\SourceCode;
 use Phpactor\Rpc\Response\ReplaceFileSourceResponse;
